pyproject.toml
[project]
name = "rl-bia-bestpractices"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gymnasium>=1.2.1",
    "pyyaml>=6.0.3",
    "swig>=4.3.1.post0",
    "tensorboard>=2.20.0",
    "torch>=2.8.0",
]
